Inline TfsHelper.Vs2012Base into TfsHelper.VS2015
In df660032 ("Remove support for VS2012 & VS2013", 2019-05-21) the support
for VS2012 and VS2013 was removed. With this, the class TfsHelperVs2012Base
no  longer serves a purpose as the only remaining class deriving from it is
TfsHelper.VS2015.

Push it down into TfsHelper.VS2015 and remove TfsHelper.Vs2012Base.cs
This is best reviewed with `git diff --color-moved=dimmed-zebra`
