+++ src/GitTfs.Vs2015/GitTfs.Vs2015.csproj
@@ -17,9 +17,6 @@
     <Compile Include="..\GitTfs.VsCommon\Retry.cs">
       <Link>Retry.cs</Link>
     </Compile>
-    <Compile Include="..\GitTfs.VsCommon\TfsHelper.Vs2012Base.cs">
-      <Link>TfsHelper.Vs2012Base.cs</Link>
-    </Compile>
     <Compile Include="..\GitTfs.VsCommon\TfsPlugin.cs">
       <Link>TfsPlugin.cs</Link>
     </Compile>
