+++ src/GitTfs.Vs2015/TfsHelper.Vs2015.cs
@@ -1,14 +1,26 @@
 using System;
 using System.Diagnostics;
 using System.IO;
+using System.Linq;
+using System.Reflection;
+
+using GitTfs.Core.TfsInterop;
 using GitTfs.VsCommon;
+
+using Microsoft.TeamFoundation.Build.Client;
+using Microsoft.TeamFoundation.Client;
+using Microsoft.TeamFoundation.Server;
+using Microsoft.TeamFoundation.VersionControl.Client;
+
 using StructureMap;
 
 namespace GitTfs.Vs2015
 {
-    public class TfsHelper : TfsHelperVs2012Base
+    public class TfsHelper : TfsHelperBase
     {
-        protected override string TfsVersionString { get { return "14.0"; } }
+        private string vsInstallDir;
+
+        private string TfsVersionString { get { return "14.0"; } }
 
         public TfsHelper(TfsApiBridge bridge, IContainer container)
             : base(bridge, container)
@@ -23,6 +35,59 @@ namespace GitTfs.Vs2015
             Trace.TraceWarning("Checkin dialog is not supported with dotnet core version of git-tfs");
             return string.Empty;
 #endif
+        }
+        protected override bool HasWorkItems(Changeset changeset)
+        {
+            return Retry.Do(() => changeset.AssociatedWorkItems.Length > 0);
+        }
+
+        private string GetVsInstallDir()
+        {
+            if (vsInstallDir == null)
+            {
+                vsInstallDir = TryGetRegString(@"Software\WOW6432Node\Microsoft\VisualStudio\" + TfsVersionString, "InstallDir")
+                    ?? TryGetRegString(@"Software\Microsoft\VisualStudio\" + TfsVersionString, "InstallDir")
+                    ?? TryGetUserRegString(@"Software\WOW6432Node\Microsoft\WDExpress\" + TfsVersionString + "_Config", "InstallDir")
+                    ?? TryGetUserRegString(@"Software\Microsoft\WDExpress\" + TfsVersionString + "_Config", "InstallDir");
+            }
+            return vsInstallDir;
+        }
+
+        protected override IBuildDetail GetSpecificBuildFromQueuedBuild(IQueuedBuild queuedBuild, string shelvesetName)
+        {
+            var build = queuedBuild.Builds.FirstOrDefault(b => b.ShelvesetName == shelvesetName);
+            return build != null ? build : queuedBuild.Build;
+        }
+
+#pragma warning disable 618
+        private IGroupSecurityService GroupSecurityService
+        {
+            get { return GetService<IGroupSecurityService>(); }
+        }
+
+        public override IIdentity GetIdentity(string username)
+        {
+            return _bridge.Wrap<WrapperForIdentity, Identity>(Retry.Do(() => GroupSecurityService.ReadIdentity(SearchFactor.AccountName, username, QueryMembership.None)));
+        }
+
+        protected override TfsTeamProjectCollection GetTfsCredential(Uri uri)
+        {
+            return HasCredentials ?
+                new TfsTeamProjectCollection(uri, GetCredential(), new UICredentialsProvider()) :
+                new TfsTeamProjectCollection(uri, new UICredentialsProvider());
+#pragma warning restore 618
+        }
+
+        protected override Assembly LoadFromVsFolder(object sender, ResolveEventArgs args)
+        {
+            Trace.WriteLine("Looking for assembly " + args.Name + " ...");
+            string folderPath = Path.Combine(GetVsInstallDir(), "PrivateAssemblies");
+            string assemblyPath = Path.Combine(folderPath, new AssemblyName(args.Name).Name + ".dll");
+            if (File.Exists(assemblyPath) == false)
+                return null;
+            Trace.WriteLine("... loading " + args.Name + " from " + assemblyPath);
+            Assembly assembly = Assembly.LoadFrom(assemblyPath);
+            return assembly;
         }
     }
 }
